requirements.txt
# QP Generation Engine — Backend Requirements
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
psycopg2-binary==2.9.9
bcrypt==4.2.0
PyJWT==2.9.0
openai==1.51.0
pymupdf==1.24.10
sentence-transformers==3.1.1
faiss-cpu==1.8.0
scikit-learn==1.5.2
numpy==1.26.4
langchain==0.3.1
pydantic==2.9.2
